presize: Precision Based Sample Size Calculation

Bland (2009) doi:10.1136/bmj.b3985 recommended to base study sizes on the width of the confidence interval rather the power of a statistical test. The goal of 'presize' is to provide functions for such precision based sample size calculations. For a given sample size, the functions will return the precision (width of the confidence interval), and vice versa.

See also

Author

Maintainer: Alan G. Haynes alan.haynes@ctu.unibe.ch

Authors:

Armando Lenz armando.lenz@ctu.unibe.ch

Andreas Limacher andreas.limacher@ctu.unibe.ch

Other contributors:

Odile Stalder odile.stalder@ctu.unibe.ch [contributor]

Marie Roumet marie.roumet@ctu.unibe.ch [contributor]