Sample size or precision for correlation coefficient

prec_cor returns the sample size or the precision for the given pearson, spearman, or kendall correlation coefficient

prec_cor(
  r,
  n = NULL,
  conf.width = NULL,
  conf.level = 0.95,
  method = c("pearson", "kendall", "spearman"),
  ...
)

Arguments

r
Desired correlation coefficient.

n
sample size

conf.width
precision (the full width of the conficende interval)

conf.level
confidence level

method
Exactly one of pearson (default), kendall, or spearman. Methods can be abbreviated.

...
other options to uniroot (e.g. tol)

Value

Object of class "presize", a list of arguments (including the computed one) augmented with method and note elements

Details

Exactly one of the parameters n, conf.width must be passed as NULL, and that parameter is determined from the other.

Sample size or precision is calculated according to formula 2 in Bonett and Wright (2000). The use of pearson is only recommended, if \(n \ge 25\). The pearson correlation coefficient assumes bivariate normality. If the assumption of bivariate normality cannot be met, spearman or kendall should be considered.

n is rounded up to the next whole number using ceiling.

uniroot is used to solve n.

References

Bonett DG, and Wright TA (2000) Sample size requirements for estimating Pearson, Kendall and Spearman correlations Psychometrika 65:23-28. doi:10.1007/BF02294183