Sample size or precision for Cohen's kappa

prec_kappa returns the sample size or the precision for the provided Cohen's kappa coefficient.

prec_kappa(
  kappa,
  n = NULL,
  raters = 2,
  n_category = 2,
  props,
  conf.width = NULL,
  conf.level = 0.95
)

Arguments

kappa
expected value of Cohen's kappa.

n
sample size.

raters
number of raters (maximum of 6).

n_category
number of categories of outcomes (maximum of 5).

props
expected proportions of each outcome (should have length n_category).

conf.width
precision (the full width of the confidence interval).

conf.level
confidence level.

Value

Object of class "presize", a list of arguments (including the computed one) augmented with method and note elements.

Details

This function wraps the FixedN and CI functions in the kappaSize package. The FixedN functions in kappaSize return a one sided confidence interval. The values that are passed to kappaSize ensure that two-sided confidence intervals are returned, although we assume that confidence intervals are symmetrical.

See also

Examples

prec_kappa,Sample size or precision for Cohen's kappa,Sample size or precision for Cohen's kappa — prec_kappa,Source: R/correlation_measures.R,prec_kappa.Rd,character(0),prec_kappa,character(0),list(title = "Description", contents = "
prec_kappa returns the sample size or the precision for the provided Cohen's kappa coefficient.
"),list(description = "prec_kappa returns the sample size or the precision for the provided Cohen's kappa coefficient."),prec_kappa(
  kappa,
  n = NULL,
  raters = 2,
  n_category = 2,
  props,
  conf.width = NULL,
  conf.level = 0.95
),list(list(name = "kappa", description = "
expected value of Cohen's kappa.
"), list(name = "n", description = "
sample size.
"), list(name = "raters", description = "
number of raters (maximum of 6).
"), list(name = "n_category", description = "
number of categories of outcomes (maximum of 5).
"), list(name = "props", description = "
expected proportions of each outcome (should have length\nn_category).
"), list(name = "conf.width", description = "
precision (the full width of the confidence interval).
"), 
    list(name = "conf.level", description = "
confidence level.
")),TRUE,
# precision based on sample size
#   two categories with proportions of 30 and 70\%, four raters
prec_kappa(kappa = .5, n = 200, raters = 4, n_category = 2, props = c(.3,.7))

#> 
#>      precision for Cohen's kappa 
#> 
#>   kappa   n   lwr   upr conf.width conf.level
#> 1   0.5 200 0.425 0.575       0.15       0.95

# sample size to get a given precision
prec_kappa(kappa = .5, conf.width = .15, raters = 4, n_category = 2,
           props = c(.3,.7))

#> 
#>      sample size for Cohen's kappa 
#> 
#>   kappa   n   lwr   upr conf.width conf.level
#> 1   0.5 198 0.425 0.575       0.15       0.95

# as above, but with two scenarios for kappa
prec_kappa(kappa = c(.5, .75), conf.width = .15, raters = 4, n_category = 2,
           props = c(.3,.7))

#> 
#>      sample size for Cohen's kappa 
#> 
#>   kappa   n   lwr   upr conf.width conf.level
#> 1  0.50 198 0.425 0.575       0.15       0.95
#> 2  0.75 155 0.675 0.825       0.15       0.95

prec_kappa(kappa = c(.5, .75), conf.width = c(.15, 0.3), raters = 4,
           n_category = 2, props = c(.3,.7))

#> 
#>      sample size for Cohen's kappa 
#> 
#>   kappa   n   lwr   upr conf.width conf.level
#> 1  0.50 198 0.425 0.575       0.15       0.95
#> 2  0.75  44 0.600 0.900       0.30       0.95

,list(list(title = "Value", contents = "
Object of class \"presize\", a list of arguments (including the\n  computed one) augmented with method and note elements.
", slug = "value"), list(title = "Details", contents = "
This function wraps the FixedN and CI functions in the\nkappaSize package.\nThe FixedN functions in kappaSize return a one sided confidence\ninterval. The values that are passed to kappaSize ensure that two-sided\nconfidence intervals are returned, although we assume that confidence intervals\nare symmetrical.
", 
    slug = "details"), list(title = "See also", contents = "

", slug = "see-also")),2021,list(name = "presize", version = "0.2.3", authors = "Armando Lenz, Alan G. Haynes, Andreas Limacher"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "presize"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  Changelog\n
", right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),FALSE