Sample size or precision for a mean difference

prec_meandiff returns the sample size or the precision for the provided mean difference and standard deviations.

prec_meandiff(
  delta,
  sd1,
  sd2 = sd1,
  n1 = NULL,
  r = 1,
  conf.width = NULL,
  conf.level = 0.95,
  variance = c("equal", "unequal"),
  ...
)

Arguments

delta
difference in means between the two groups.

sd1
standard deviation in group 1.

sd2
standard deviation in group 2.

n1
number of patients in group 1.

r
allocation ratio (relative size of group 2 and group 1 (n2 / n1)).

conf.width
precision (the full width of the confidence interval).

conf.level
confidence level.

variance
equal (default) or unequal variance.

...
other options to uniroot (e.g. tol)

Value

Object of class "presize", a list of arguments (including the computed one) augmented with method and note elements.

Details

Exactly one of the parameters n or conf.width must be passed as NULL, and that parameter is determined from the other.

Examples

prec_meandiff,Sample size or precision for a mean difference,Sample size or precision for a mean difference — prec_meandiff,Source: R/differences.R,prec_meandiff.Rd,character(0),prec_meandiff,character(0),list(title = "Description", contents = "
prec_meandiff returns the sample size or the precision for the\nprovided mean difference and standard deviations.
"),list(description = "prec_meandiff returns the sample size or the precision for the\nprovided mean difference and standard deviations."),prec_meandiff(
  delta,
  sd1,
  sd2 = sd1,
  n1 = NULL,
  r = 1,
  conf.width = NULL,
  conf.level = 0.95,
  variance = c("equal", "unequal"),
  ...
),list(list(name = "delta", description = "
difference in means between the two groups.
"), list(name = "sd1", description = "
standard deviation in group 1.
"), list(name = "sd2", description = "
standard deviation in group 2.
"), list(name = "n1", description = "
number of patients in group 1.
"), list(name = "r", description = "
allocation ratio (relative size of group 2 and group 1 (n2 / n1)).
"), list(name = "conf.width", description = "
precision (the full width of the confidence interval).
"), 
    list(name = "conf.level", description = "
confidence level.
"), list(name = "variance", description = "
equal (default) or unequal variance.
"), list(name = "...", description = "
other options to uniroot (e.g. tol)
")),TRUE,
prec_meandiff(delta = 5, sd1 = 2.5, n1 = 20, var = "equal")

#> 
#>      precision for mean difference with equal variance 
#> 
#>   delta sd1 sd2 n1 n2 conf.width conf.level      lwr      upr
#> 1     5 2.5 2.5 20 20   3.044614       0.95 3.477693 6.522307

prec_meandiff(delta = 5, sd1 = 2.5, conf.width = 3, var = "equal")

#> 
#>      sample size for mean difference with equal variance 
#> 
#>   delta sd1 sd2       n1       n2 conf.width conf.level lwr upr
#> 1     5 2.5 2.5 20.61861 20.61861          3       0.95 3.5 6.5
,list(list(title = "Value", contents = "
Object of class \"presize\", a list of arguments (including the\n  computed one) augmented with method and note elements.
", slug = "value"), list(title = "Details", contents = "
Exactly one of the parameters n or conf.width must be passed as NULL,\nand that parameter is determined from the other.
", slug = "details")),2021,list(name = "presize", version = "0.2.1", authors = "Armando Lenz, Alan G. Haynes, Andreas Limacher"),list(destination = "dev", mode = "release", version_label = "default", version_tooltip = "Released version", in_dev = FALSE),list(root = "../", title = "presize"),list(),list(type = "default", left = "
\n  \n    \n     \n  \n
\n
\n  Reference\n
\n
\n  Changelog\n
", right = "
\n  \n    \n     \n  \n
"),list(.present = TRUE),list(version = "1.6.1"),FALSE